# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  common_models.FilterOperator:
    enum:
    - IS
    - IS_NOT
    - IS_ONE_OF_TERMS
    - IS_ONE_OF_TERMS_OR
    - IS_LESS_THAN_OR_EQUALS
    - IS_GREATER_THAN
    - EXIST
    - DOES_NOT_EXIST
    - NOT_CONTAINS
    - CONTAIN
    - DOES_NOT_CONTAIN
    - CONTAIN_ONE_OF
    - DOES_NOT_CONTAIN_ONE_OF
    - IS_ONE_OF
    - IS_NOT_ONE_OF
    - IS_BETWEEN
    - IS_NOT_BETWEEN
    - IS_IN_FIELDS
    - IS_NOT_IN_FIELDS
    - ENDS_WITH
    - NOT_ENDS_WITH
    - START_WITH
    - NOT_START_WITH
    type: string
    x-enum-varnames:
    - OpEquals
    - OpNotEquals
    - OpIn
    - OpInOr
    - OpLessOrEq
    - OpGreater
    - OpExists
    - OpNotExists
    - OpNotContains
    - OpContain
    - OpDoesNotContain
    - OpContainOneOf
    - OpDoesNotContainOneOf
    - OpIsOneOf
    - OpIsNotOneOf
    - OpIsBetween
    - OpIsNotBetween
    - OpIsInFields
    - OpIsNotInFields
    - OpEndsWith
    - OpNotEndsWith
    - OpStartWith
    - OpNotStartWith
  common_models.FilterType:
    properties:
      field:
        type: string
      operator:
        $ref: '#/definitions/common_models.FilterOperator'
      value: {}
    type: object
...
  domain.FailedIndex:
    properties:
      index_name:
        type: string
      index_uuid:
        type: string
      reason:
        type: string
    type: object
...
  domain.PolicySettings:
    properties:
      deleteAfter:
        type: string
      snapshotActive:
        type: boolean
    type: object
...
  domain.UpdateManagedIndexPolicyResponse:
    properties:
      failed_indices:
        items:
          $ref: '#/definitions/domain.FailedIndex'
        type: array
      failures:
        type: boolean
      updated_indices:
        type: integer
    type: object
...
  dto.IndexInfo:
    properties:
      docs.count:
        type: string
      health:
        type: string
      index:
        type: string
      status:
        type: string
      store.size:
        type: string
    type: object
...
  dto.IndexPropertyType:
    properties:
      name:
        type: string
      type:
        type: string
    type: object
...
  dto.PropertyValuesWithCountRequest:
    properties:
      field:
        type: string
      filters:
        items:
          $ref: '#/definitions/common_models.FilterType'
        type: array
      index:
        type: string
      orderByCount:
        type: boolean
      sortAsc:
        type: boolean
      top:
        type: integer
    required:
    - field
    - index
    type: object
...
info:
  contact: {}
  description: UTMStack backend API — manages IAM, modules, flows, incidents, compliance
    and audit data.
  title: UTMStack Backend Service
  version: "1.0"
paths:
...
  /index-policy/policy:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PolicySettings'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get index policy settings (returns empty body if no policy configured)
      tags:
      - Index Policy
    put:
      consumes:
      - application/json
      parameters:
      - description: Policy settings (snapshotActive bool, deleteAfter string)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.PolicySettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UpdateManagedIndexPolicyResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update index policy settings
      tags:
      - Index Policy
...
  /opensearch/index/all:
    get:
      description: Returns a paginated list of OpenSearch indices, optionally filtered
        by pattern.
      parameters:
      - description: Include system indices (starting with '.')
        in: query
        name: includeSystemIndex
        type: boolean
      - description: Filter by index name pattern
        in: query
        name: pattern
        type: string
      - description: Page number (1-based)
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: Total number of matching indices
              type: string
          schema:
            items:
              $ref: '#/definitions/dto.IndexInfo'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all indices
      tags:
      - OpenSearch
  /opensearch/index/delete-index:
    post:
      consumes:
      - application/json
      description: Deletes one or more OpenSearch indices by name.
      parameters:
      - description: List of index names to delete
        in: body
        name: indices
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "200":
          description: Deleted
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete indices
      tags:
      - OpenSearch
  /opensearch/index/properties:
    get:
      description: Returns a flattened list of field name + type for the given index
        pattern.
      parameters:
      - description: Index pattern
        in: query
        name: indexPattern
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.IndexPropertyType'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get index properties
      tags:
      - OpenSearch
  /opensearch/property/values:
    get:
      description: Returns all distinct values for a keyword field across the given
        index pattern.
      parameters:
      - description: Keyword field name
        in: query
        name: keyword
        required: true
        type: string
      - description: Index pattern
        in: query
        name: indexPattern
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get field values
      tags:
      - OpenSearch
  /opensearch/property/values-with-count:
    post:
      consumes:
      - application/json
      description: Returns a map of field value → document count for the given index
        and field.
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.PropertyValuesWithCountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get field values with document count
      tags:
      - OpenSearch
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"